Migrate fresh + rollback works again

+++ database/migrations/2019_03_15_113418_add_status_column_to_buildings_table.php
@@ -25,8 +25,10 @@ class AddStatusColumnToBuildingsTable extends Migration
      */
     public function down()
     {
-        Schema::table('buildings', function (Blueprint $table) {
-            $table->dropColumn('status');
-        });
+        if (Schema::hasColumn('buildings', 'status')) {
+            Schema::table('buildings', function (Blueprint $table) {
+                $table->dropColumn('status');
+            });
+        }
     }
 }
